requirements.txt
absl-py==2.3.0
asgiref==3.8.1
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.2
flatbuffers==25.2.10
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.73.0
h5py==3.14.0
idna==3.10
keras==2.11.0
libclang==18.1.1
Markdown==3.8
MarkupSafe==3.0.2
numpy==1.24.4
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==25.0
pillow==11.2.1
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
